backend/requirements.txt
langchain==0.3.25
langchain-community==0.3.23
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-chroma==0.2.4
langchain-text-splitters==0.3.8
chromadb==1.0.9
pypdf==5.1.0
fastapi==0.115.9
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
tiktoken==0.8.0
requests==2.32.3
fastembed==0.4.2